Initial push of a match-3
First pass at match 3 game.

Currently in.  When 4 or more of the same color is connected delete the objects.

move the player left or right at the bottom of the screen to select the bottom most block.

Hit j to grab and place a block

Have the blocks move from the top of the screen to the bottom of the screen.

End the game when 8 lines of blocks are on the screen.

ToDo:

Make the blocks fall back to the next closest block.
Add some sounds to the sample.
Improve the graphics to look more complete
Scoring system.